+++ src/numerical/ode.rs
@@ -1,12 +1,16 @@
 use structure::matrix::*;
 use structure::vector::*;
 use std::ops::{Add, Sub, Mul, Div};
+use std::convert::Into;
 pub use self::ODEMethod::*;
+use numerical::runge_kutta::rk4;
+use numerical::bdf::bdf1;
+use structure::dual::*;
 
 #[derive(Debug, Copy, Clone)]
 pub enum ODEMethod {
     RK4,
-    BDF1,
+    BDF1(f64),
 }
 
 /// ODE Solver
@@ -15,9 +19,55 @@ pub enum ODEMethod {
 /// * `f = f(t, y)`
 /// * `init_value = y_start`
 /// * `param_range = (t_start, t_end)`
-pub fn solve<F, G>(f: F, init_value: Vec<G>, param_range: (f64, f64), step: f64, method: ODEMethod) -> Matrix
-    where F: Fn(Vec<G>) -> Vec<G> + Copy,
-          G: Add<f64> + Sub<f64> + Mul<f64> + Div<f64> + Copy,
+///
+/// # Type
+/// `solve: (F, Vec<f64>, (T, T), f64, ODEMethod) -> Matrix where Fn(Dual, Vec<Dual>) -> Vec<Dual> + Copy, T: Into<f64> + Copy`
+///
+/// # Examples
+/// ```
+/// extern crate peroxide;
+/// use peroxide::*;
+///
+/// let init_val = c!(2, 1);
+/// let result = solve(lotka_volterra, init_val, (0, 10), 1e-3, RK4);
+/// result.print();
+///
+/// fn lotka_volterra(t: Dual, xs: Vec<Dual>) -> Vec<Dual> {
+///     let a = 4.;
+///     let c = 1.;
+///
+///     let x = xs[0];
+///     let y = xs[1];
+///
+///     vec![
+///         a * (x - x*y),
+///         -c * (y - x*y)
+///     ]
+/// }
+/// ```
+pub fn solve<F, T>(f: F, init_value: Vec<f64>, param_range: (T, T), step: f64, method: ODEMethod) -> Matrix
+    where F: Fn(Dual, Vec<Dual>) -> Vec<Dual> + Copy,
+          T: Into<f64> + Copy
 {
-    unimplemented!()
+    let t_start = param_range.0.into();
+    let t_end = param_range.1.into();
+    let num = ((t_end - t_start) / step).round() as usize;
+
+    match method {
+        RK4 => {
+            let g = |t: f64, xs: Vec<f64>| f(dual(t, 0.), merge_dual(xs.clone(), vec![0f64;xs.len()])).values();
+            let result = rk4(
+                t_start,
+                init_value.into_iter().map(|p| p.into()).collect::<Vec<f64>>(),
+                g,
+                step,
+                num,
+            );
+            result
+        },
+        BDF1(rtol) => {
+            let result = bdf1(t_start, init_value, f, step, rtol, num);
+            result
+        }
+    }
 }